+++ src/react-components/src/__tests__/components/Label-test.tsx
@@ -1,0 +1,22 @@
+
+import { expect } from "chai";
+import * as Enzyme from "enzyme";
+import * as React from "react";
+import * as Mocha from "mocha";
+
+import Label from "../../../src/components/01-atoms/Forms/Label";
+
+describe("Label Test", () => {
+  let wrapper: Enzyme.ShallowWrapper<{}, {}>;
+  it("Renders a <label> when passed text as children", () => {
+  });
+
+  it("Renders a <label> when passed text as prop", () => {
+  });
+
+  it("Renders a <label> when passed element as children", () => {
+  });
+
+  it("Throws error when neither text nor children are passed", () => {
+  });
+});