+++ src/react-components/src/__tests__/components/Pagination-test.tsx
@@ -1,0 +1,13 @@
+
+import { expect } from "chai";
+import * as Enzyme from "enzyme";
+import * as React from "react";
+import * as Mocha from "mocha";
+
+import Pagination from "../../../src/components/02-molecules/Menus/Pagination/Pagination";
+
+describe("Pagination Test", () => {
+  let wrapper: Enzyme.ShallowWrapper<{}, {}>;
+  it("Renders two buttons and a Select", () => {
+  });
+});