+++ src/react-components/src/components/01-atoms/Forms/Label.stories.tsx
@@ -1,11 +1,10 @@
 import * as React from "react";
 
 import Label from "./Label";
-import { action } from "@storybook/addon-actions";
 
 export default {
   title: "Label",
   component: Label,
 };
 
-export const label = () => <Label referenceId="should-never-be-used-alone"><div>Lonely</div></Label>;
+export const label = () => <Label id="label-id" referenceId="should-never-be-used-alone"><div>Lonely</div></Label>;