+++ src/react-components/src/components/02-molecules/Menus/Pagination/Pagination.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from "react";
+import { action } from "@storybook/addon-actions";
+
+import Pagination from "./Pagination";
+
+export default {
+  title: "Pagination",
+  component: Pagination,
+};
+
+
+export const pagination = () => <Pagination
+totalPages={10}
+previousPageHandler={action("goToPrevious")}
+nextPageHandler={action("goToNext")}
+onSelectChange={action("selectChange")}
+onSelectBlur={action("selectBlur")}/>;