+++ src/react-components/src/components/02-molecules/Menus/Pagination/Pagination.tsx
@@ -2,13 +2,58 @@
 
 import * as React from "react";
 import bem from "../../../../utils/bem";
-import { Button } from "src/react-components/src";
+import Button from "../../../01-atoms/Button/Button";
+import Dropdown from "../../../02-molecules/Dropdown/Dropdown";
 
 export interface PaginationProps {
-
+  totalPages: number;
+  previousPageHandler: (event: React.MouseEvent) => void;
+  nextPageHandler: (event: React.MouseEvent) => void;
+  onSelectChange: (event: React.MouseEvent) => void;
+  onSelectBlur: (event: React.MouseEvent) => void;
 }
 
 export default function Pagination(props: PaginationProps) {
-  <Button></Button>
-  
+  const { totalPages,
+    previousPageHandler, nextPageHandler,
+    onSelectChange, onSelectBlur } = props;
+
+  const pagination__base_class = "pagination";
+  let buttonPrevProps = {
+    id: "pagination-previous",
+    callback: previousPageHandler,
+    content: "Previous",
+    type: "outline",
+    iconPosition: "left",
+    iconName: "arrow-xsmall",
+    iconModifiers: ["left"],
+    iconDecorative: true,
+  };
+
+  let buttonNextProps = {
+    id: "pagination-next",
+    callback: nextPageHandler,
+    content: "Next",
+    type: "outline",
+    iconPosition: "right",
+    iconName: "arrow-xsmall",
+    iconModifiers: ["right"],
+    iconDecorative: true,
+  };
+
+  let dropdownProps = {
+    dropdownId: "pagination-dropdown",
+    isRequired: false,
+    labelPosition: "left",
+    labelId: "pagination-select-label",
+    labelText: "Page",
+    dropdownOptions: [...Array(totalPages).keys()].map(i =>  (i + 1).toString() + " of " + totalPages.toString()),
+    onSelectChange: onSelectChange,
+    onSelectBlur: onSelectBlur
+  };
+
+  return (
+    <nav aria-label="Pagination" className={bem(pagination__base_class)}>
+      <Button {...buttonPrevProps}/><Dropdown {...dropdownProps}></Dropdown><Button {...buttonNextProps}/>
+  </nav>);
 }